+++ src/shapes/cone.h
@@ -51,7 +51,7 @@ public:
     bool IntersectP(const Ray &ray) const;
     float Area() const;
 
-    size_t toGPU(Metadata* meta, void* data) const;
+    uint32_t toGPU(Metadata* meta, void* data) const;
 protected:
     // Cone Private Data
     float radius, height, phiMax;
