+++ src/shapes/cylinder.cpp
@@ -211,7 +211,7 @@ Point Cylinder::Sample(float u1, float u2, Normal *Ns) const {
     return (*ObjectToWorld)(p);
 }
 
-size_t Cylinder::toGPU(Metadata* meta, void* data) const {
+uint32_t Cylinder::toGPU(Metadata* meta, void* data) const {
 	Severe("Cylinder not implemented for GPU");
 	return 0;
 }
