+++ src/shapes/disk.cpp
@@ -156,7 +156,7 @@ Point Disk::Sample(float u1, float u2, Normal *Ns) const {
     return (*ObjectToWorld)(p);
 }
 
-size_t Disk::toGPU(Metadata* meta, void* data) const {
+uint32_t Disk::toGPU(Metadata* meta, void* data) const {
 	Severe("Disk not implemented for GPU");
 	return 0;
 }
