+++ src/shapes/hyperboloid.cpp
@@ -253,7 +253,7 @@ float Hyperboloid::Area() const {
               2.f*p1.z*p2.z - p2.z*p2.z));
 }
 
-size_t Hyperboloid::toGPU(Metadata* meta, void* data) const {
+uint32_t Hyperboloid::toGPU(Metadata* meta, void* data) const {
 	Severe("Hyperboloid not implemented for GPU");
 	return 0;
 }
